# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API by Housni BENABID",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/brands": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "List brands",
                "description": "List brands endpoint",
                "operationId": "brand.index",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/brand/{slug}": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Show brand",
                "description": "Show brand endpoint",
                "operationId": "brand.show",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Brand slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Brands"
                ],
                "summary": "Update brand",
                "description": "Create brand endpoint",
                "operationId": "brand.update",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Brand slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create brand request body",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "The title of the brand",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "slug": {
                                        "description": "The slug of the brand, auto-generated if not provided",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Brands"
                ],
                "summary": "Update brand",
                "description": "Create brand endpoint",
                "operationId": "brand.destroy",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Brand slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/brand/create": {
            "post": {
                "tags": [
                    "Brands"
                ],
                "summary": "Create brand",
                "description": "Create brand endpoint",
                "operationId": "brand.create",
                "requestBody": {
                    "description": "Create brand request body",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "The title of the brand",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "slug": {
                                        "description": "The slug of the brand, auto-generated if not provided",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/main/blog": {
            "get": {
                "tags": [
                    "Main"
                ],
                "summary": "List blogposts",
                "description": "List blogposts endpoint",
                "operationId": "blog.index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/main/blog/{slug}": {
            "get": {
                "tags": [
                    "Main"
                ],
                "summary": "Show blogpost",
                "description": "Show blogpost endpoint",
                "operationId": "blog.show",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Blogpost slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/": {
            "get": {
                "tags": [
                    "User endpoint"
                ],
                "summary": "Show user details",
                "description": "Show user details endpoint",
                "operationId": "user.show",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User endpoint"
                ],
                "summary": "Delete authenticated user",
                "description": "Delete authenticated user endpoint and receive the confirmation",
                "operationId": "user.destroy",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}